+++ vm/src/types/instance_definitions/builtins_instance_def.rs
@@ -139,9 +139,9 @@ impl BuiltinsInstanceDef {
             keccak: Some(KeccakInstanceDef::new(Some(2048), vec![200; 8])),
             poseidon: Some(PoseidonInstanceDef::new(Some(256))),
             #[cfg(feature = "mod_builtin")]
-            add_mod: Some(ModInstanceDef::default()),
+            add_mod: Some(ModInstanceDef::new(Some(128), 1, 96)),
             #[cfg(feature = "mod_builtin")]
-            mul_mod: Some(ModInstanceDef::default()),
+            mul_mod: Some(ModInstanceDef::new(Some(256), 1, 96)),
             #[cfg(not(feature = "mod_builtin"))]
             add_mod: None,
             #[cfg(not(feature = "mod_builtin"))]
@@ -159,13 +159,7 @@ impl BuiltinsInstanceDef {
             ec_op: Some(EcOpInstanceDef::default()),
             keccak: None,
             poseidon: None,
-            #[cfg(feature = "mod_builtin")]
-            add_mod: Some(ModInstanceDef::new(Some(32), 8)),
-            #[cfg(feature = "mod_builtin")]
-            mul_mod: Some(ModInstanceDef::new(Some(32), 8)),
-            #[cfg(not(feature = "mod_builtin"))]
             add_mod: None,
-            #[cfg(not(feature = "mod_builtin"))]
             mul_mod: None,
         }
     }
@@ -180,7 +174,13 @@ impl BuiltinsInstanceDef {
             ec_op: Some(EcOpInstanceDef::new(None)),
             keccak: None,
             poseidon: None,
+            #[cfg(feature = "mod_builtin")]
+            add_mod: Some(ModInstanceDef::new(None, 1, 96)),
+            #[cfg(feature = "mod_builtin")]
+            mul_mod: Some(ModInstanceDef::new(None, 1, 96)),
+            #[cfg(not(feature = "mod_builtin"))]
             add_mod: None,
+            #[cfg(not(feature = "mod_builtin"))]
             mul_mod: None,
         }
     }